Get rid of sync demos from PollyDemos
