+++ PollyDemos/Demo02_WaitAndRetryNTimes.cs
@@ -1,19 +1,21 @@
+using PollyDemos.Helpers;
 using PollyDemos.OutputHelpers;
 
-namespace PollyDemos.Async;
+namespace PollyDemos;
 
 /// <summary>
-/// Demonstrates the Retry strategy coming into action.
+/// Demonstrates the Retry strategy with delays between retry attempts.
 /// Loops through a series of HTTP requests, keeping track of each requested
 /// item and reporting server failures when encountering exceptions.
 ///
-/// Observations: There's no wait among these retries. It can be appropriate sometimes.
-/// In this case, no wait hasn't given underlying system time to recover, so calls still fail despite retries.
+/// Observations: We now have waits among the retries.
+/// In this case, still not enough wait - or not enough retries - for the underlying system to have recovered.
+/// So we still fail some calls.
 /// </summary>
-public class AsyncDemo01_RetryNTimes : AsyncDemo
+public class Demo02_WaitAndRetryNTimes : DemoBase
 {
     public override string Description =>
-        "This demo demonstrates a first Retry.  It retries three times, immediately.";
+        "Compared to previous demo, this adds waits between the retries. Not always enough wait to ensure success, tho.";
 
     public override async Task ExecuteAsync(CancellationToken cancellationToken, IProgress<DemoProgress> progress)
     {
@@ -26,23 +28,20 @@ public class AsyncDemo01_RetryNTimes : AsyncDemo
 
         PrintHeader(progress);
 
-        // Define our strategy:
         var strategy = new ResiliencePipelineBuilder().AddRetry(new()
         {
             ShouldHandle = new PredicateBuilder().Handle<Exception>(),
-            MaxRetryAttempts = 3, // Retry up to 3 times
+            MaxRetryAttempts = 3,
+            Delay = TimeSpan.FromMilliseconds(200), // Wait between each try
             OnRetry = args =>
             {
-                // Due to how we have defined ShouldHandle, this delegate is called only if an exception occurred.
-                // Note the ! sign (null-forgiving operator) at the end of the command.
-                var exception = args.Outcome.Exception!; // The Exception property is nullable
-
-                // Tell the user what happened
+                var exception = args.Outcome.Exception!;
                 progress.Report(ProgressWithMessage($"Strategy logging: {exception.Message}", Color.Yellow));
                 Retries++;
                 return default;
             }
         }).Build();
+
 
         var client = new HttpClient();
         var internalCancel = false;
@@ -53,12 +52,8 @@ public class AsyncDemo01_RetryNTimes : AsyncDemo
 
             try
             {
-                // Retry the following call according to the strategy.
-                // The cancellationToken passed in to ExecuteAsync() enables the strategy to cancel retries when the token is signalled.
                 await strategy.ExecuteAsync(async token =>
                 {
-                    // This code is executed within the strategy
-
                     var responseBody = await IssueRequestAndProcessResponseAsync(client, token);
                     progress.Report(ProgressWithMessage($"Response : {responseBody}", Color.Green));
                     EventualSuccesses++;
