+++ PollyDemos/Demo05_WaitAndRetryWithExponentialBackoff.cs
@@ -1,24 +1,27 @@
+using PollyDemos.Helpers;
 using PollyDemos.OutputHelpers;
 
-namespace PollyDemos.Sync;
+namespace PollyDemos;
 
 /// <summary>
-/// Demonstrates the Retry strategy with delays between retry attempts.
+/// Demonstrates Retry strategy with calculated retry delays to back off.
 /// Loops through a series of HTTP requests, keeping track of each requested
 /// item and reporting server failures when encountering exceptions.
 ///
-/// Observations: We no longer have to guess how many retries are enough.
-/// All calls still succeed!  Yay!
-/// But we're still hammering that underlying server with retries.
-/// Imagine if lots of clients were doing that simultaneously
-///  - could just increase load on an already-struggling server!
+/// Observations: All calls still succeed!  Yay!
+/// But we didn't hammer the underlying server so hard - we backed off.
+/// That's healthier for it, if it might be struggling ...
+/// ... and if a lot of clients might be doing this simultaneously.
+///
+/// ... What if the underlying system was totally down tho?
+/// ... Keeping trying forever would be counterproductive (so, see Demo06)
 /// </summary>
-public class Demo04_WaitAndRetryForever : SyncDemo
+public class Demo05_WaitAndRetryWithExponentialBackoff : DemoBase
 {
     public override string Description =>
-        "This demo also retries enough to always ensure success. But we haven't had to 'guess' how many retries were necessary. We just said: wait-and-retry-forever.";
+        "This demonstrates exponential back-off. We have enough retries to ensure success. But we don't hammer the server so hard: we increase the delay between each try.";
 
-    public override void Execute(CancellationToken cancellationToken, IProgress<DemoProgress> progress)
+    public override async Task ExecuteAsync(CancellationToken cancellationToken, IProgress<DemoProgress> progress)
     {
         ArgumentNullException.ThrowIfNull(progress);
 
@@ -32,12 +35,13 @@ public class Demo04_WaitAndRetryForever : SyncDemo
         var strategy = new ResiliencePipelineBuilder().AddRetry(new()
         {
             ShouldHandle = new PredicateBuilder().Handle<Exception>(),
-            MaxRetryAttempts = int.MaxValue, // Retry indefinitely
-            Delay = TimeSpan.FromMilliseconds(200),
+            MaxRetryAttempts = 6, // We could also retry indefinitely by using int.MaxValue
+            BackoffType = DelayBackoffType.Exponential, // Back off: 1s, 2s, 4s, 8s, ... + jitter
             OnRetry = args =>
             {
                 var exception = args.Outcome.Exception!;
                 progress.Report(ProgressWithMessage($"Strategy logging: {exception.Message}", Color.Yellow));
+                progress.Report(ProgressWithMessage($" ... automatically delaying for {args.RetryDelay.TotalMilliseconds}ms.", Color.Yellow));
                 Retries++;
                 return default;
             }
@@ -52,9 +56,9 @@ public class Demo04_WaitAndRetryForever : SyncDemo
 
             try
             {
-                strategy.Execute(token =>
+                await strategy.ExecuteAsync(async token =>
                 {
-                    var responseBody = IssueRequestAndProcessResponse(client, token);
+                    var responseBody = await IssueRequestAndProcessResponseAsync(client, token);
                     progress.Report(ProgressWithMessage($"Response : {responseBody}", Color.Green));
                     EventualSuccesses++;
 
@@ -66,7 +70,7 @@ public class Demo04_WaitAndRetryForever : SyncDemo
                 EventualFailures++;
             }
 
-            Thread.Sleep(500);
+            await Task.Delay(TimeSpan.FromSeconds(0.5), cancellationToken);
             internalCancel = ShouldTerminateByKeyPress();
         }
     }
